+++ src/main/java/com/aetherteam/aether/command/EternalDayCommand.java
@@ -2,7 +2,7 @@ package com.aetherteam.aether.command;
 
 import com.aetherteam.aether.AetherConfig;
 import com.aetherteam.aether.attachment.AetherDataAttachments;
-import com.aetherteam.aether.attachment.INBTSynchable;
+import com.aetherteam.aether.attachment.AttachmentSyncable;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.BoolArgumentType;
 import net.minecraft.commands.CommandSourceStack;
@@ -32,7 +32,7 @@ public class EternalDayCommand {
             data.setEternalDay(value);
             data.updateEternalDay(level); // Syncs to client.
             if (AetherConfig.SERVER.sync_aether_time.get()) {
-                data.setSynched(-1, INBTSynchable.Direction.DIMENSION, "setShouldWait", true, level);
+                data.setSynched(-1, AttachmentSyncable.Direction.DIMENSION, "setShouldWait", true, level);
             }
             source.sendSuccess(() -> Component.translatable("commands.aether.capability.time.eternal_day.set", value), true);
         }