+++ src/main/java/com/aetherteam/aether/event/hooks/PlayerAttachmentSyncHooks.java
@@ -1,7 +1,7 @@
 package com.aetherteam.aether.event.hooks;
 
 import com.aetherteam.aether.attachment.AetherDataAttachments;
-import com.aetherteam.aether.attachment.INBTSynchable;
+import com.aetherteam.aether.attachment.AttachmentSyncable;
 import net.minecraft.world.entity.player.Player;
 
 public final class PlayerAttachmentSyncHooks {
@@ -10,7 +10,7 @@ public final class PlayerAttachmentSyncHooks {
 
     public static void syncPlayerAttachment(Player player) {
         if (!player.level().isClientSide()) {
-            player.getAttachedOrCreate(AetherDataAttachments.AETHER_PLAYER).forceSync(player.getId(), INBTSynchable.Direction.CLIENT);
+            player.getAttachedOrCreate(AetherDataAttachments.AETHER_PLAYER).forceSync(player.getId(), AttachmentSyncable.Direction.CLIENT);
         }
     }
 }