+++ src/main/java/com/aetherteam/aether/mixin/mixins/common/AbstractArrowMixin.java
@@ -1,7 +1,7 @@
 package com.aetherteam.aether.mixin.mixins.common;
 
 import com.aetherteam.aether.attachment.AetherDataAttachments;
-import com.aetherteam.aether.attachment.INBTSynchable;
+import com.aetherteam.aether.attachment.AttachmentSyncable;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.projectile.arrow.AbstractArrow;
@@ -32,7 +32,7 @@ public abstract class AbstractArrowMixin {
         if (arrow.hasAttached(AetherDataAttachments.PHOENIX_ARROW)) {
             var attachment = arrow.getAttachedOrCreate(AetherDataAttachments.PHOENIX_ARROW);
             if (attachment.isPhoenixArrow() && !arrow.level().isClientSide()) {
-                attachment.setSynched(arrow.getId(), INBTSynchable.Direction.CLIENT, "setPhoenixArrow", true); // Sync Phoenix Arrow variable to client.
+                attachment.setSynched(arrow.getId(), AttachmentSyncable.Direction.CLIENT, "setPhoenixArrow", true); // Sync Phoenix Arrow variable to client.
                 if (this.isInGround()) { // Spawn less particles when the arrow is in the ground.
                     if (this.inGroundTime % 5 == 0) {
                         this.spawnParticles(arrow);