+++ src/main/java/com/aetherteam/aether/network/packet/SyncEntityPacket.java
@@ -1,6 +1,6 @@
 package com.aetherteam.aether.network.packet;
 
-import com.aetherteam.aether.attachment.INBTSynchable;
+import com.aetherteam.aether.attachment.AttachmentSyncable;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
@@ -9,15 +9,15 @@ import org.apache.commons.lang3.tuple.Triple;
 
 import javax.annotation.Nullable;
 
-public abstract class SyncEntityPacket<T extends INBTSynchable> extends SyncPacket<T> {
+public abstract class SyncEntityPacket<T extends AttachmentSyncable> extends SyncPacket<T> {
     private final int entityID;
 
-    protected SyncEntityPacket(int entityID, String key, INBTSynchable.Type valueType, Object value) {
+    protected SyncEntityPacket(int entityID, String key, AttachmentSyncable.Type valueType, Object value) {
         super(key, valueType, value);
         this.entityID = entityID;
     }
 
-    protected SyncEntityPacket(Quartet<Integer, String, INBTSynchable.Type, Object> values) {
+    protected SyncEntityPacket(Quartet<Integer, String, AttachmentSyncable.Type, Object> values) {
         this(values.getA(), values.getB(), values.getC(), values.getD());
     }
 
@@ -31,13 +31,13 @@ public abstract class SyncEntityPacket<T extends INBTSynchable> extends SyncPack
         super.write(buf);
     }
 
-    public static Quartet<Integer, String, INBTSynchable.Type, Object> decodeEntityValues(RegistryFriendlyByteBuf buf) {
+    public static Quartet<Integer, String, AttachmentSyncable.Type, Object> decodeEntityValues(RegistryFriendlyByteBuf buf) {
         int entityID = buf.readVarInt();
-        Triple<String, INBTSynchable.Type, Object> values = decodeValues(buf);
+        Triple<String, AttachmentSyncable.Type, Object> values = decodeValues(buf);
         return new Quartet<>(entityID, values.getLeft(), values.getMiddle(), values.getRight());
     }
 
-    public static <T extends INBTSynchable> void execute(SyncEntityPacket<T> payload, @Nullable Player player) {
+    public static <T extends AttachmentSyncable> void execute(SyncEntityPacket<T> payload, @Nullable Player player) {
         if (player == null) {
             return;
         }